Fix converged files and streamed replacements.
When files were replaced and streaming was in use a error would be
thrown from the logging code if the file was ignored - which happens
when the content has converged with that already on disk (e.g. because
an earlier mirror run was interrupted).

Secondly, when files had converged, we would end up deleting the file
rather than leaving it alone, due to a bug in replay() that would
affect all users, not just streaming.